% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biv.rec.sim.R
\name{biv.rec.sim}
\alias{biv.rec.sim}
\title{Deprecated: Use simBivRec}
\usage{
biv.rec.sim(nsize, beta1, beta2, tau_c, set)
}
\arguments{
\item{nsize}{Sample size which refers to the number of subjects in the data set where each subject could have multiple episodes of events.}

\item{beta1}{True coefficients for Type I gap times in the accelerated failure time model (AFT).}

\item{beta2}{True coefficients for Type II gap times in the accelerated failure time model (AFT).}

\item{tau_c}{Maximum support of censoring time. It can take values as follows:
\itemize{
\item \code{tau_c=63} (default): corresponds to a 15\% censoring rate for each scenario in Tables 1 and 2 of Lee, Huang, Xu, Luo (2018).
\item \code{tau_c=30}: corresponds to a 30\% censoring rate for each scenario in Tables 1 and 2 of Lee, Huang, Xu, Luo (2018).
}}

\item{set}{Simulation setting based on scenarios outlined in Tables 1 and 2 in Lee, Huang, Xu, Luo (2018). Choose 1.1 (default) for scenario 1 with \eqn{\rho=1} in the covariance matrix of the frailty vector, 1.2 for scenario 1 with \eqn{\rho=0.5}, 1.3 for scenario 1 with \eqn{\rho=0} and 2.0 for scenario 2. Default is 1.1.}
}
\value{
Data frame with the alternating recurrent event data and one continuous and one binary covariate.
}
\description{
Deprecated function from the previous version. Use \code{simBivRec}.
}
\references{
Lee CH, Huang CY, Xu G, Luo X. (2018). Semiparametric regression analysis for alternating recurrent event data. Statistics in Medicine, 37: 996-1008.
\url{https://doi.org/10.1002/sim.7563}
}
