% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivrecFormula.R
\name{bivrecFormula}
\alias{bivrecFormula}
\title{Create a bivrecFormula object.}
\usage{
bivrecFormula(x)
}
\arguments{
\item{x}{A formula with a bivrecSurv object on the left of a '~' operator as response, and the covariate(s) on the right.}
}
\value{
A bivrecFormula object.
}
\description{
Create a bivrecFormula object.
}
