% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivrecNP.R, R/is.R
\name{bivrecNP}
\alias{bivrecNP}
\alias{is_bivrecNP}
\title{Non-parametric analysis that obtains marginal survival, conditional cdf and joint cdf}
\usage{
bivrecNP(x, CI, ai, u1, u2, conditional, given.interval)

is_bivrecNP(x)
}
\arguments{
\item{x}{an object to be checked with is_bivrecNP}

\item{CI}{ffsf}

\item{ai}{dd}

\item{u1}{ffd}

\item{u2}{dd}

\item{conditional}{dd}

\item{given.interval}{af}
}
\description{
Non-parametric analysis that obtains marginal survival, conditional cdf and joint cdf
}
