% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chang_univariate.R
\name{chang_univariate}
\alias{chang_univariate}
\title{A Function for univariate fits using semiparametric regression method on a biv.rec object}
\usage{
chang_univariate(new_data, cov_names, SE)
}
\arguments{
\item{new_data}{An object that has been reformatted for fit using the biv.rec.reformat() function. Passed from biv.rec.fit().}

\item{cov_names}{A string with the name of the covariate. Passed from biv.rec.fit().}

\item{CI}{Passed from biv.rec.fit().}
}
\value{
A list with estimates, SE and variance-covariance matrix.
}
\description{
This function fits the model using Chang's Method given one covariate. Called from biv.rec.fit(). No user interface.
}
\keyword{internal}
