% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_vcov_confint.R
\name{confint.bivrecReg}
\alias{confint.bivrecReg}
\title{Confidence Interval for the Coefficients of a Semiparametric Regression Fit}
\usage{
\method{confint}{bivrecReg}(object, parm, level, ...)
}
\arguments{
\item{object}{A \verb{bivrecReg} object.}

\item{parm}{The parameters for which to run confidence interval. Default gets CI for all the covariates in the model.}

\item{level}{Significance level. Default is 0.95.}

\item{...}{Additional parameters if needed.}
}
\description{
This function obtains the confidence interval for the coefficients of a semiparametric regression fit.
}
