% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_vcov_confint.R
\name{confint}
\alias{confint}
\title{confint}
\usage{
confint(object, parm, level)
}
\arguments{
\item{object}{A bivrecReg object obtained by using bivrecReg() function}

\item{parm}{The parameters for which to run confidence interval. Default is giving CI for all the covariates in the model.}

\item{level}{Significance level. Example: 0.99 for a 99\% confidence interval. Default is 0.95.}
}
\description{
Obtain confidence interval for exponentiated coefficients of semi-parametric regression fit using bivrecReg
}
