% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.sim.R
\name{data.sim}
\alias{data.sim}
\title{A Data Simulation Function}
\usage{
data.sim(nsize, beta1, beta2, cr, sg2, set)
}
\arguments{
\item{nsize}{sample size}

\item{beta1}{true coefficients for first gap time}

\item{beta2}{true coefficients for second gap time}

\item{cr}{maximum support of censoring time}

\item{sg2}{variance of frailty}

\item{set}{optional simulation setting. Choose 1.1 (default) for $rho=1$ in covariance matrix, 1.2 for $rho=0.5$ in covariance matrix or 2.1 for $rho=0$ in covariance matrix.}
}
\value{
Data frame with alternating recurrent event data and two covariates
}
\description{
This function simulates a series of alternating recurrent events based on a linear combination of two covariates.
}
\examples{
data.sim(nsize=300, beta1=c(0.5,0.5), beta2=c(0,-0.5), cr=63, sg2=0.5, set=1.1)

}
\keyword{data.sim,}
\keyword{simulation}
