% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head.R
\name{head.bivrecNP}
\alias{head.bivrecNP}
\title{Print first five elements of the Joint CDF, Marginal Survival, and Conditional CDF from a bivrecNP object}
\usage{
\method{head}{bivrecNP}(x, ...)
}
\arguments{
\item{x}{A bivrecNP object}

\item{...}{additional parameters if needed}
}
\description{
Print first five elements of the Joint CDF, Marginal Survival, and Conditional CDF from a bivrecNP object
}
