% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isBivRec.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Function to check object class wholenumber}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\description{
Function to check object class wholenumber
}
\keyword{internal}
