% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leeall_multivariate.R
\name{leeall_multivariate}
\alias{leeall_multivariate}
\title{A Function for multivariate fits using semiparametric regression method on a bivrecSurv object}
\usage{
leeall_multivariate(response, amat, cov_names, SE)
}
\arguments{
\item{response}{Passed from bivrecReg().}

\item{amat}{Passed from bivrecReg().}

\item{cov_names}{Passed from bivrecReg().}

\item{SE}{Passed from bivrecReg()}
}
\value{
A list with estimates, SE and variance-covariance matrix.
}
\description{
This function fits the semiparametric model given one  covariate. Called from bivrecReg(). No user interface.
}
\keyword{internal}
