% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marg.surv.plot.R
\name{marg.surv.plot}
\alias{marg.surv.plot}
\title{Plot Marginal Survival after non-parametric analysis}
\usage{
marg.surv.plot(bivrec.nonparam.result, CI)
}
\arguments{
\item{bivrec.nonparam.result}{List with marginal.survival, formula and data. Passed from biv.rec.np()}

\item{CI}{Confidence level for CI. Passed from biv.rec.np()}
}
\value{
A plot of marginal survival vs. first gap time with confidence interval.
}
\description{
This function plots marginal survival for recurrent event data.  Called from biv.rec.np(). No user interface.
}
\keyword{internal}
