% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparam.cdf.R
\name{nonparam.cdf}
\alias{nonparam.cdf}
\title{A Function for non-parametric analysis on a biv.rec object}
\usage{
nonparam.cdf(fit_data, u, ai)
}
\arguments{
\item{fit_data}{An object that has been reformatted using the biv.rec.reformat() function. Passed from biv.rec.fit() function.}

\item{u}{Passed from biv.rec.fit() function.}

\item{ai}{Passed from biv.rec.fit() function.}
}
\value{
A dataframe with the joint CDF
}
\description{
This function calculates the joint CDF for bivariate recurrent events. Called from biv.rec.fit(). No user interface.
}
\seealso{
\code{\link{biv.rec.fit}}
}
\keyword{internal}
