% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparam.conditional.R
\name{nonparam.conditional}
\alias{nonparam.conditional}
\title{A Function for additional non-parametric analysis of bivariate recurrent event data.}
\usage{
nonparam.conditional(bivrec.nonparam.result, given.interval)
}
\arguments{
\item{bivrec.nonparam.result}{is a list obtained from running biv.rec.fit with method="Non-Parametric",}

\item{given.interval}{is a 1x2 vector indicating an interval for the first gap time to estimate the cdf of second gap time given this interval. If interval given is c(v1, v2) the  function calculates P(yij <= w | v1 <= xij <= v2).
The given values v1 and v2 must be in the range of gap times in the estimated marginal survival (valid times are given in the Time column of the results for marginal survival).}
}
\value{
A data frame with the conditional CDF for the given an interval of the first gap time and corresponding plot.
}
\description{
This function calculates the conditional cdf after estimation of joint cdf and marginal survival.
}
\keyword{internal}
