% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparam.marginal.R
\name{nonparam.marginal}
\alias{nonparam.marginal}
\title{A Function for non-parametric analysis on a biv.rec object}
\usage{
nonparam.marginal(fit_data)
}
\arguments{
\item{fit_data}{An object that has been reformatted using the biv.rec.reformat() function. Passed from biv.rec.np().}
}
\value{
A data frame with marginal survival
}
\description{
This function calculates the marginal survival for bivariate recurrent events. Called from biv.rec.np(). No user interface.
}
\keyword{internal}
