% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparam_cdf.R
\name{nonparam_cdf}
\alias{nonparam_cdf}
\title{A Function for non-parametric analysis on a biv.rec object for joint cdf}
\usage{
nonparam_cdf(fit_data, u, ai, CI)
}
\arguments{
\item{fit_data}{Passed from bivrecNP().}

\item{u}{Passed from bivrecNP().}

\item{ai}{Passed from bivrecNP().}

\item{CI}{Passed from bivrecNP().}
}
\value{
A dataframe with the joint CDF
}
\description{
This function calculates the joint CDF for bivariate recurrent events. Called from bivrecNP(). No user interface.
}
\keyword{internal}
