% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNP.R
\name{plot.bivrecNP}
\alias{plot.bivrecNP}
\title{Plot graphs (joint cdf, marginal and conditional cdf (if user wants)) after non-parametric analysis}
\usage{
\method{plot}{bivrecNP}(x)
}
\arguments{
\item{x}{an object of class \code{bivrecNP}}
}
\description{
This function plots joint cdf, marginal survival and conditional cdf for recurrent event data.  Called from biv.rec.np(). No user interface.
}
