% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{Plot Bivariate Alternating Recurrent Series}
\usage{
plot(x, y, data, ...)
}
\arguments{
\item{x}{either coordinates for a plot or an object of class \code{bivrecSurv}, \code{bivrecReg}, \code{bivrecNP} or a formula.}

\item{y}{the y coordinates of points for a plot if coordinates were used in x, not required otherwise}

\item{data}{argument only when x is a formula. Should inidicate the data frame that contains categorical covariates and/or vectors to create the response indicate in the given formula.}

\item{...}{arguments to be passed to graphical methods as needed.
\itemize{
 \item main: String with plot title (default is no title)
 \item xlab: String with label for horizontal axis (default is "Gap Times")
 \item ylab: String with label for vertical axis (default is "Individual")
 \item type1: String to label type 1 gap times (default is "Type 1")
 \item type2: String to label type 2 gap times (default is "Type 2")
}}
}
\description{
This function plots bivariate recurrent event gap times in different format
}
\examples{
# Simulate bivariate alternating recurrent event data
library(BivRec)
set.seed(1234)
bivrec_data <- simulate(nsize=150, beta1=c(0.5,0.5), beta2=c(0,-0.5), tau_c=63, set=1.1)

plot(with(bivrec_data, bivrecSurv(id, epi, xij, yij, d1, d2)), main="Example")
plot(bivrecSurv(id, epi, xij, yij, d1, d2) ~ a1 + a2, data = bivrec_data, main="Example")

}
