% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bivrecNP.R
\name{plotCond}
\alias{plotCond}
\title{Plot Conditional CDF after Nonparametric Analysis}
\usage{
plotCond(object, type)
}
\arguments{
\item{object}{An object of \verb{bivrecNP} class where the analysis has specified \verb{conditional = TRUE}.}

\item{type}{Optional string to label the Type II gap time. Default is "Type II".}
}
\description{
This function plots the conditional cdf for a \verb{bivrecNP} object.
}
