% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bivrecNP.R
\name{plotJoint}
\alias{plotJoint}
\title{Plot Joint CDF after Nonparametric Analysis}
\usage{
plotJoint(object, type)
}
\arguments{
\item{object}{A \verb{bivrecNP} object.}

\item{type}{Optional vector of strings to label Type I and Type II gap times. Default is c("Type I", "Type II").}
}
\description{
This function plots the joint cdf for a \verb{bivrecNP} object.
}
