% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bivrecNP.R
\name{plotMarg}
\alias{plotMarg}
\title{Plot Marginal Survival after Nonparametric Analysis}
\usage{
plotMarg(object, type)
}
\arguments{
\item{object}{An object of \verb{bivrecNP} class.}

\item{type}{Optional string to label the Type I gap time. Default is "Type I".}
}
\description{
This function plots the marginal survival for a \verb{bivrecNP} object.
}
