% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMethods.R
\name{print.bivrecReg}
\alias{print.bivrecReg}
\title{Print a Semiparametric Regression Fit}
\usage{
\method{print}{bivrecReg}(x, ...)
}
\arguments{
\item{x}{An object of class \verb{bivrecReg}.}

\item{...}{Additional parameters if needed.}
}
\description{
This function prints an object of class \verb{bivrecReg}.
}
