% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize}
\alias{summarize}
\title{summarize}
\usage{
summarize(object)
}
\arguments{
\item{object}{a bivrecReg object obtained by using bivrecReg() function}
}
\description{
Summarize bivrecReg object
}
