% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.bivrecReg}
\alias{summary.bivrecReg}
\title{Summary of a Semiparametric Regression Fit}
\usage{
\method{summary}{bivrecReg}(object, ...)
}
\arguments{
\item{object}{A \verb{bivrecReg} object.}

\item{...}{Additional parameters if needed.}
}
\description{
This function provides a summary for the fit from semiparametric regression analysis.
}
