% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_vcov_confint.R
\name{vcov.bivrecReg}
\alias{vcov.bivrecReg}
\title{Variance-Covariance Matrix from a Semiparametric Regression Fit}
\usage{
\method{vcov}{bivrecReg}(object, ...)
}
\arguments{
\item{object}{A \verb{bivrecReg} object.}

\item{...}{Additional parameters if needed.}
}
\description{
This function extracts the variance-covariance matrix from the fit of a semiparametric regression analysis.
}
