% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_vcov_confint.R
\name{vcov}
\alias{vcov}
\title{vcov}
\usage{
vcov(object)
}
\arguments{
\item{object}{A bivrecReg object obtained by using bivrecReg() function}
}
\description{
Obtain variance-covariance matrix from semi-parametric regression fit using bivrecReg
}
