% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biv.rec.fit.R
\name{biv.rec.fit}
\alias{biv.rec.fit}
\title{Analysis of Bivariate Recurrent Data}
\usage{
biv.rec.fit(formula, data, method, CI, ai, u1, u2, condgx)
}
\arguments{
\item{formula}{Formula indicating model to fit in the form: \strong{ID + xij + yij + episode + c_indicatorY + c_indicatorX ~ covariates}.
where
\strong{ID:} is a vector of subjects' unique identifier.
\strong{xij:} is a vector with the lengths of time spent in event of type X for individual i in episode j.
\strong{yij:} is a vector with the lengths of time spent in event of type Y for individual i in episode j.
\strong{episode:} is a vector indicating the observation or episode j for a subject i.
\strong{c_indicatorY:} is an vector of indicators with values of 0 for the last episode for subject i and 1 otherwise. A subject with only one episode will have one 0.
\strong{c_indicatorX:} is an optional vector of indicators with values of 0 if the last episode for subject i occurred for event of type X or 1 otherwise. A subject with only one episode could have only a 1 (if he was censored at event Y) or a 0 (if he was censored at event X).
\strong{covariates:} the names of the covariates in the form covariate_1 + ... + covariate_N. If using Non-parametric method replaces the names with a 1.}

\item{data}{A data frame that includes all the vectors/covariates listed in the formula}

\item{method}{A string indicating which method to use to estimate effect of covariates (and corresponding 95\% CI when specified) or cdf. Options are: "Lee.all" (default), "Chang", or "Non-Parametric". For details on methods see vignette.}

\item{CI}{A logical variable. If TRUE a 95\% Confidence Interval for the effect estimates is calculated. Default is FALSE.}

\item{ai}{A real non-negative function of censoring time to be used as weights in the Non-Parametric Method. Current options are 1 such that ai = f(censoring times) = 1 or the default of 2 such that ai = f(censoring times) = censoring times.}

\item{u1}{A vector or single number to be used for Estimation of joint CDF P(xij <= u1, yij <= u2) in the Non-Parametric Method. (see vignette for further details)}

\item{u2}{A vector or single number to be used for Estimation of joint CDF P(xij <= u1, yij <= u2) in the Non-Parametric Method. (see vignette for further details)}

\item{condgx}{Automatically passed - not for user}
}
\value{
A list with dataframes of results and some of the parameters used to produce them.
}
\description{
This function allows the user to:
1) Evaluate covariate effects on two alternating states/gap times.
2) Find the joint cdf, marginal survival and conditional cdf for two alternating states/gap times.
}
\examples{
\dontrun{
library(BivRec)
set.seed(1234)
sim.data <- data.sim(nsize=300, beta1=c(0.5,0.5), beta2=c(0,-0.5), cr=63, sg2=0.5, set=1.1)
multiple.lee <- biv.rec.fit(id + xij + yij + epi + d2 + d1 ~ a1 + a2,
                data=sim.data, method="Lee.all", CI=FALSE)
multiple.lee
nonpar <- biv.rec.fit(id + xij + yij + epi + d2 + d1 ~ 1,
          data=sim.data, method="Non-Parametric", ai=1)
nonpar
}

}
\keyword{biv.rec.fit}
