% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biv.rec.fit.R
\name{biv.rec.fit}
\alias{biv.rec.fit}
\title{Semi-Parametric Accelerated Failure Time Analysis of Bivariate Recurrent Data}
\usage{
biv.rec.fit(formula, data, method, CI)
}
\arguments{
\item{formula}{Formula in the form \strong{ID + xij + yij + episode + c_indicatorY + c_indicatorX ~ covariates} indicating model to fit.
 where
\itemize{
  \item ID: is a vector of subjects' unique identifier.
  \item xij: is a vector with the lengths of time spent in event of type X for individual i in episode j.
  \item yij: is a vector with the lengths of time spent in event of type Y for individual i in episode j.
  \item episode: is a vector indicating the observation or episode j for a subject i.
  \item c_indicatorY: is an vector of indicators with values of 0 for the last episode for subject i and 1 otherwise. A subject with only one episode will have one 0.
  \item c_indicatorX: is an optional vector of indicators with values of 0 if the last episode for subject i occurred for event of type X or 1 otherwise. A subject with only one episode could have only a 1 (if he was censored at event Y) or a 0 (if he was censored at event X).
  \item covariates: the names of the covariates in the form covariate_1 + ... + covariate_N. If using Non-parametric method replaces the names with a 1.
}}

\item{data}{A data frame that includes all the vectors/covariates listed in the formula}

\item{method}{A string indicating which method to use to estimate effect of covariates (and corresponding 95\% CI when specified). Options are: "Lee.all" (default) or "Chang". For details on methods see vignette.}

\item{CI}{Level for confidence interval between 0.50 and 0.99, 0.99 would give 99\% CI. Default is 0.95.}
}
\value{
A BivRec list object containing:
\itemize{
  \item \strong{covariate.effects:} Data Frame summarizing effects of the covariates including point estimate, standard error and confidence interval.
  \item \strong{formula:} The formula used to specify components of bivariate recurrent response and covariates.
}
}
\description{
This function allows the user to evaluate covariate effects on two alternating recurrent events/gap times using the methods outlined in: Lee CH, Huang C-Y, Xu G, Luo X. Semiparametric regression analysis for alternating recurrent event data. Statistics in Medicine. 2018;37:996-1008. \url{https://doi.org/10.1002/sim.7563}
}
\examples{
\dontrun{
library(BivRec)
set.seed(1234)
sim.data <- data.sim(nsize=150, beta1=c(0.5,0.5), beta2=c(0,-0.5), cr=63, sg2=0.5, set=1.1)
multiple.lee <- biv.rec.fit(formula = id + xij + yij + epi + d2 + d1 ~ a1 + a2,
                data=sim.data, method="Lee.all", CI=TRUE)
multiple.lee
}

}
\keyword{biv.rec.fit}
