% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biv.rec.np.R
\name{biv.rec.np}
\alias{biv.rec.np}
\title{Non-Parametric Analysis of Bivariate Recurrent Event Data}
\usage{
biv.rec.np(formula, data, ai, u1, u2, marginalplot, conditional, given.interval)
}
\arguments{
\item{formula}{Formula in the form: \strong{ID + xij + yij + episode + c_indicatorY + c_indicatorX ~ 1}.
where
\strong{ID:} is a vector of subjects' unique identifier.
\strong{xij:} is a vector with the lengths of time spent in event of type X for individual i in episode j.
\strong{yij:} is a vector with the lengths of time spent in event of type Y for individual i in episode j.
\strong{episode:} is a vector indicating the observation or episode j for a subject i.
\strong{c_indicatorY:} is an vector of indicators with values of 0 for the last episode for subject i and 1 otherwise. A subject with only one episode will have one 0.
\strong{c_indicatorX:} is an optional vector of indicators with values of 0 if the last episode for subject i occurred for event of type X or 1 otherwise. A subject with only one episode could have only a 1 (if he was censored at event Y) or a 0 (if he was censored at event X).}

\item{data}{A data frame that includes all the vectors/covariates listed in the formula}

\item{ai}{A real non-negative function of censoring time to be used as weights in the Non-Parametric Method. Current options are 1 such that ai = f(censoring times) = 1 or the default of 2 such that ai = f(censoring times) = censoring times.}

\item{u1}{A vector or single number to be used for Estimation of joint CDF P(xij <= u1, yij <= u2) in the Non-Parametric Method. (see vignette for further details)}

\item{u2}{A vector or single number to be used for Estimation of joint CDF P(xij <= u1, yij <= u2) in the Non-Parametric Method. (see vignette for further details)}

\item{marginalplot}{Logical. If TRUE (default) this function will plot the marginal survival for the first gap time with 95\% confidence interval.}

\item{conditional}{Logical. If TRUE will calculate the conditional cdf for the second gap time given an interval for the first gap time. Must provide given.interval.}

\item{given.interval}{is a 1x2 vector indicating an interval for the first gap time to estimate the cdf of second gap time given this interval. If given.interval = c(v1, v2) the function calculates P(yij <= u2 | v1 <= xij <= v2).
The given values v1 and v2 must be in the range of gap times in the estimated marginal survival (valid times are given in the Time column of the results for marginal survival).}
}
\value{
A list with dataframes of results and some of the parameters used to produce them.
}
\description{
This function allows the user to find the joint cdf, marginal survival and conditional cdf for two alternating states/gap times.
}
\examples{
\dontrun{
library(BivRec)
set.seed(1234)
sim.data <- data.sim(nsize=300, beta1=c(0.5,0.5), beta2=c(0,-0.5), cr=63, sg2=0.5, set=1.1)
nonpar.result <- biv.rec.fit(id + xij + yij + epi + d2 + d1 ~ 1,
          data=sim.data, ai=1)
nonpar.result$cdf
}

}
\keyword{biv.rec.np}
