% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biv.rec.plot.R
\name{biv.rec.plot}
\alias{biv.rec.plot}
\title{Bivariate Recurrent Response Plotting}
\usage{
biv.rec.plot(formula, data)
}
\arguments{
\item{formula}{Formula of the form \strong{ID ~ xij + yij}.
\itemize{
  \item ID: is a vector of unique subject identifiers.
  \item xij: is a vector with the length of time subject i spends in event of type 1 during observation j.
  \item yij: is a vector with the length of time subject i spends in event of type 2 during observation j.
}}

\item{data}{A data frame that contains all the vectors listed in the formula}
}
\description{
This function plots recurrent event data.
}
\examples{
\dontrun{
library(BivRec)
set.seed(1234)
sim.biv.rec <- data.sim(300, beta1=c(0.5,0.5), beta2=c(0,-0.5), cr=63, sg2=0.5, set=1.1)
biv.rec.plot(formula = id ~ xij + yij, data = sim.biv.rec)
}

}
