% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biv.rec.plot.R
\name{biv.rec.plot}
\alias{biv.rec.plot}
\title{Deprecated: Use \code{plot.bivrecSurv}}
\usage{
biv.rec.plot(formula, data)
}
\arguments{
\item{formula}{Formula of the form \strong{ID + episode ~ xij + yij}.
\itemize{
  \item ID: A vector of subjects' unique identifier which can be numeric or character.
  \item episode: A vector indicating the episode of the bivariate alternating gap time pairs.
  \item xij: A vector with the lengths of the Type I gap times.
  \item yij: A vector with the lengths of the Type II gap times.
}}

\item{data}{A data frame that contains all the vectors listed in the formula}
}
\description{
Deprecated function from previous version. Use \code{plot.bivrecSurv}.
}
