% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biv.rec.sim.R
\name{biv.rec.sim}
\alias{biv.rec.sim}
\title{Bivariate Recurrent Response and Covariate Data Simulation}
\usage{
biv.rec.sim(nsize, beta1, beta2, cr, sg2, set)
}
\arguments{
\item{nsize}{sample size}

\item{beta1}{true coefficients for first gap time}

\item{beta2}{true coefficients for second gap time}

\item{cr}{maximum support of censoring time}

\item{sg2}{variance of frailty}

\item{set}{Simulation setting based on scenerios outlined by Lee et all. Choose 1.1 (default) for scenerio 1 with rho=1 in covariance matrix, 1.2 for scenerio 1 with rho=0.5, 1.3 for scenario 1 with rho=0 and 2.0 for scenario 2.}
}
\value{
Data frame with alternating recurrent event data and two covariates
}
\description{
This function simulates a series of alternating recurrent events based on simulations studies in: Lee CH, Huang C-Y, Xu G, Luo X. Semiparametric regression analysis for alternating recurrent event data. Statistics in Medicine. 2018;37:996-1008. \url{https://doi.org/10.1002/sim.7563}
}
\examples{
biv.rec.sim(nsize=300, beta1=c(0.5,0.5), beta2=c(0,-0.5), cr=63, sg2=0.5, set=1.1)

}
\keyword{biv.rec.sim,}
\keyword{simulation}
