% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivrecSurv.R
\name{bivrecSurv}
\alias{bivrecSurv}
\title{Create a Bivariate Recurrent Survival Object}
\usage{
bivrecSurv(id, episode, xij, yij, d1, d2)
}
\arguments{
\item{id}{Vector of subject's unique identifier (i).}

\item{episode}{Vector indicating the observation or episode (j) for a subject (i). This will determine order of events for each subject.}

\item{xij}{Vector with the lengths of time spent in event of type X for individual i in episode j.}

\item{yij}{vector with the lengths of time spent in event of type Y for individual i in episode j.}

\item{d1}{Vector of indicators, with values of 0 if the last episode for subject i occurred for event of type X or 1 otherwise. A subject with only one episode could have either one 1 (if he was censored at event Y) or one 0 (if he was censored at event X). A subject with censoring in event Y will have a vector of 1's.}

\item{d2}{Vector of indicators, with values of 0 for the last episode for subject i or 1 otherwise. A subject with only one episode will have one 0.}
}
\value{
A bivrecSurv object ready to be used as the response for analysis using bivrecReg or bivrecNP.
}
\description{
This function creates a bivariate recurrent survival object to be used as a response variable in a model formula.
}
\examples{
set.seed(1234)
sim_data <- simBivRec(nsize=150, beta1=c(0.5,0.5), beta2=c(0,-0.5))
bivrecsurv_data <- with(sim_data, bivrecSurv(id, epi, xij, yij, d1, d2))

}
