% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivrecNP.R
\name{bivrecNP}
\alias{bivrecNP}
\title{nonpar.result <- biv.rec.np(formula = id + epi + xij + yij + d1 + d2 ~ 1,
          data=dat, ai=1, u1 = c(2, 5, 10, 20), u2 = c(1, 5, 10, 15),
          conditional = FALSE, given.interval=c(0, 10), jointplot=FALSE,
          marginalplot = FALSE, condiplot = FALSE)}
\usage{
bivrecNP(x, CI, ai, u1, u2, conditional, given.interval)
}
\description{
nonpar.result2 <- biv.rec.np(formula = id + epi + xij + yij + d1 + d2 ~ 1,
          data=dat, ai=1, u1 = c(2, 5, 10, 20), u2 = c(1, 5, 10, 15),
          conditional = TRUE, given.interval=c(0, 10), jointplot=FALSE,
          marginalplot = FALSE, condiplot = FALSE)
}
