% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivrecNP.R
\name{bivrecNP}
\alias{bivrecNP}
\title{Non-Parametric Analysis of Bivariate Alternating Recurrent Event Gap Time Data}
\usage{
bivrecNP(response, ai, u1, u2, CI, conditional, given.interval)
}
\arguments{
\item{response}{A response object of the \code{bivrecSurv} class.}

\item{ai}{A real non-negative function of censoring time. See details.}

\item{u1}{A vector or single number to be used for estimation of joint cdf \eqn{P(X0 \le u1, Y0 \le u2)} in the non-parametric method.}

\item{u2}{A vector or single number to be used for estimation of joint cdf \eqn{P(X0 \le u1, Y0 \le u2)} in the non-parametric method.}

\item{CI}{The level for confidence intervals for joint cdf plot, marginal plot and conditional cdf. Must be between 0.50 and 0.99, where 0.99 would give 99\% CI. Default is 0.95.}

\item{conditional}{A logical value. If TRUE, this function will calculate the conditional cdf for the type II gap time given an interval of the type I gap time and the bootstrap standard error and confidence interval at the specified confidence level. Default is FALSE.}

\item{given.interval}{A vector c(v1, v2) that must be specified if conditional = TRUE. The vector indicates an interval for the type I gap time to use for estimation of the cdf of the type II gap time given this interval.
If given.interval = c(v1, v2), the function calculates \eqn{P(Y0 \le y | v1 \le X0 \le v2)}. The given values v1 and v2 must be in the range of gap times in the estimated marginal survival.}
}
\description{
This function allows the user to obtain the joint, conditional and marginal cumulative distribution functions.
See details for the estimation methods provided.
}
\details{
\strong{ai} indicates a real non-negative function of censoring times to be used as weights in the non-parametric method. This variable can take on values of 1 or 2 which indicate:
\itemize{
\item 1: the weights are simply 1 for all subjects \eqn{a(C_i) = 1} (default).
\item 2: the weight for each subject is his/her censoring time \eqn{a(C_i) = C_i}.
}
}
\examples{

library(BivRec)

# Simulate bivariate alternating recurrent event data
set.seed(1234)
sim_data <- simulate(nsize=150, beta1=c(0.5,0.5), beta2=c(0,-0.5), tau_c=63, set=1.1)
npresult <- bivrecNP(response = with(sim_data, bivrecSurv(id, epi, xij, yij, d1, d2)),
                     ai=1, u1 = c(2, 5, 10, 20), u2 = c(1, 5, 10, 15))
#' plot(npresult)
\dontrun{
#This is an example with longer runtime (it runs the conditional graph)
 npresult2 <- bivrecNP(response = with(sim_data, bivrecSurv(id, epi, xij, yij, d1, d2)),
                     ai=1, u1 = c(2, 5, 10, 20), u2 = c(1, 5, 10, 15),
 plot(npresult2)
}
}
