% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivrecReg.R
\name{bivrecReg}
\alias{bivrecReg}
\title{Semi-Parametric Accelerated Failure Time Analysis of Bivariate Alternating Recurrent Event Gap Time Data}
\usage{
bivrecReg(formula, data, method)
}
\arguments{
\item{formula}{A formula with a bivrecSurv object as response.}

\item{data}{A data frame that includes all the covariates listed in the formula.}

\item{method}{A string indicating which method to use to estimate effects of the covariates. See details.}
}
\value{
A bivrecReg object containing:
\itemize{
  \item \strong{covariate.effects:} A data frame summarizing effects of the covariates including the point estimate, standard error and confidence interval.
  \item \strong{formula:} The formula used to specify the model.
}
}
\description{
This function allows the user to evaluate covariate effects on two alternating recurrent events gap times (referred as type I and type II gap times)
under the assumption that the two gap times follow accelerated failure time (AFT) models. See details for the estimation methods provided.
}
\details{
Two different estimation methods are available:
\itemize{
\item  method = "Lee.et.al" (default) is a U-statistics-based smooth estimating function approach. See Lee CH, Huang C-Y, Xu G, Luo X (2017) for further details.
\item  method = "Chang" is a rank-based estimating function approach.  See Chang (2004) for further details.
Note that following the Chang method, the variances of the estimated regression coefficients are approximated using the resampling techniques developed by Parzen, Wei and Ying (1994).
This approximation requires extensive computing time for a relatively small sample size. In addition, using the Chang method does not guarantee convergence for the estimation of the coefficients.
}
}
\examples{
library(BivRec)
# Simulate bivariate alternating recurrent event data
set.seed(1234)
bivrec_data <- simulate(nsize=150, beta1=c(0.5,0.5), beta2=c(0,-0.5), tau_c=63, set=1.1)
# Apply Lee C, Huang CY, Xu G, Luo X (2017) method using one covariate
lee_reg <- bivrecReg(formula = bivrecSurv(id, epi, xij, yij, d1, d2) ~ a1 + a2,
                    data = bivrec_data, method="Lee.et.al")
summary(lee_reg)
plot(lee_reg)
\dontrun{

#To apply Chang (2004) method use method="Chang". This is an example with longer runtime.

library(BivRec)
# Simulate bivariate alternating recurrent event data
set.seed(1234)
bivrec_data <- simulate(nsize=150, beta1=c(0.5,0.5), beta2=c(0,-0.5), tau_c=63, set=1.1)

# Apply Chang (2004) method using multiple covariates
# and 99\% confidence intervals.
fit_chang <- bivrecReg(bivrecSurv(id, epi, xij, yij, d1, d2) ~ a1 + a2,
                    data = bivrec_data, method = "Chang")
summary(fit_chang)
}

}
\references{
\enumerate{
\item Chang S-H. (2004). Estimating marginal effects in accelerated failure time models for serial sojourn times among repeated events. Lifetime Data Analysis, 10: 175-190.
\url{https://doi.org/10.1023/B:LIDA.0000030202.20842.c9}

\item Lee C, Huang CY, Xu G, Luo X (2017). Semiparametric regression analysis for alternating recurrent event data. Statistics in Medicine, 37: 996-1008.
\url{https://doi.org/10.1002/sim.7563}

\item Parzen MI, Wei LJ, Ying Z (1994). A resampling method based on pivotal estimating functions. Biometrika, 81: 341-350.
\url{https://doi.org/10.1093/biomet/81.2.341}
}
}
\keyword{bivrecReg}
