% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chang_multivariate.R
\name{chang.multivariate}
\alias{chang.multivariate}
\title{A Function for multivariate fits using semiparametric regression method on a biv.rec object}
\usage{
chang.multivariate(new_data, cov_names, SE)
}
\arguments{
\item{new_data}{An object that has been reformatted for fit using the biv.rec.reformat() function. Passed from biv.rec.fit().}

\item{cov_names}{A vector with the names of the covariates. Passed from biv.rec.fit().}

\item{SE}{Passed from biv.rec.fit().}
}
\value{
A list with estimates, SE and variance-covariance matrix.
}
\description{
This function fits the model using Chang's Method given multiple  covariates. Called from biv.rec.fit(). No user interface.
}
\keyword{internal}
