% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_vcov_confint.R
\name{coef.bivrecReg}
\alias{coef.bivrecReg}
\title{Extract the coefficients of a semi-parametric regression fit obtained from bivrecReg}
\usage{
\method{coef}{bivrecReg}(object, ...)
}
\arguments{
\item{object}{A bivrecReg object}

\item{...}{additional parameters if needed}
}
\description{
Extract the coefficients of a semi-parametric regression fit obtained from bivrecReg
}
