% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_vcov_confint.R
\name{coef.bivrecReg}
\alias{coef.bivrecReg}
\title{Extract the Coefficients of a Semiparametric Regression Fit Obtained from a \code{bivrecReg} object}
\usage{
\method{coef}{bivrecReg}(object, ...)
}
\arguments{
\item{object}{A bivrecReg object.}

\item{...}{Additional parameters if needed.}
}
\description{
Extract the Coefficients of a Semiparametric Regression Fit Obtained from a \code{bivrecReg} object
}
