% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_vcov_confint.R
\name{coef.bivrecReg}
\alias{coef.bivrecReg}
\title{coef}
\usage{
\method{coef}{bivrecReg}(object)
}
\arguments{
\item{object}{A bivrecReg object obtained by using bivrecReg() function}
}
\description{
Obtain coefficients from semi-parametric regression fit using bivrecReg
}
