% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marg.surv.plot.R
\name{marg.surv.plot}
\alias{marg.surv.plot}
\title{Plot Marginal Survival after non-parametric analysis}
\usage{
marg.surv.plot(bivrec.nonparam.result)
}
\arguments{
\item{bivrec.nonparam.result}{is a list obtained from running biv.rec.fit with method="Non-Parametric"}
}
\value{
A plot of marginal survival vs. first time with 95\% confidence interval.
}
\description{
This function plots marginal survival for recurrent event data.
}
\examples{
\dontrun{
library(BivRec)
set.seed(1234)
sim.data <- data.sim(nsize=300, beta1=c(0.5,0.5), beta2=c(0,-0.5), cr=63, sg2=0.5, set=1.1)
nonpar.example <- biv.rec.fit(id + xij + yij + epi + d2 + d1 ~ 1,
          data=sim.data, method="Non-Parametric", ai=1)
nonpar.example
plot.marg.surv(nonpar.example)
}

}
\keyword{internal}
