% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_jointcdf.R
\name{nonparam.cdf}
\alias{nonparam.cdf}
\title{A Function for non-parametric analysis on a biv.rec object for joint cdf}
\usage{
nonparam.cdf(fit_data, u, ai, CI)
}
\arguments{
\item{fit_data}{An object that has been reformatted using the biv.rec.reformat() function. Passed from biv.rec.np().}

\item{u}{Passed from biv.rec.np().}

\item{ai}{Passed from biv.rec.np().}

\item{CI}{Passed from biv.rec.np().}
}
\value{
A dataframe with the joint CDF
}
\description{
This function calculates the joint CDF for bivariate recurrent events. Called from biv.rec.np(). No user interface.
}
\keyword{internal}
