% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bivrecNP.R
\name{plot.bivrecNP}
\alias{plot.bivrecNP}
\title{Plot Results of Nonparametric Analysis}
\usage{
\method{plot}{bivrecNP}(x, y = NULL, type = NULL, ...)
}
\arguments{
\item{x}{A \verb{bivrecNP} object.}

\item{y}{Either empty or NULL.}

\item{type}{Optional vector of strings to label Type I and Type II gap times. Default is c("Type I", "Type II").}

\item{...}{Additional arguments to be passed to graphical methods if needed.}
}
\description{
This function plots all the estimated functions (joint cdf, marginal survival and conditional cdf if \verb{conditional=TRUE} during analysis) from a \verb{bivrecNP} object in one step.
}
