% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bivrecNP.R
\name{plot.bivrecNP}
\alias{plot.bivrecNP}
\title{Plot Results of Nonparametric Analysis of Bivariate Recurrent Events}
\usage{
\method{plot}{bivrecNP}(x, y = NULL, type = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{bivrecNP}.}

\item{y}{Either empty or NULL.}

\item{type}{Optional vector of strings to label Type 1 and Type 2 gap times. Default is c("Type 1", "Type 2").}

\item{...}{Additional arguments to be passed to graphical methods if needed.}
}
\description{
This function plots all the estimated functions (joint cdf, marginal survival and/or conditional cdf) from a \code{bivrecNP} object in one step.
}
