% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bivrecNP.R
\name{plot.bivrecNP}
\alias{plot.bivrecNP}
\title{Plot Results of Nonparametric Analysis of Bivariate Recurrent Events}
\usage{
\method{plot}{bivrecNP}(x, y = NULL, type = NULL, main = NULL,
  xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{bivrecNP}.}

\item{y}{Either empty or NULL.}

\item{type}{Optional vector of strings to label type 1 and type 2 gap times. Default is c("Type 1", "Type 2").}

\item{main}{Optional string with plot title. Default is no title.}

\item{xlab}{Optional string with label for horizontal axis.}

\item{ylab}{Optional string with label for vertical axis.}

\item{...}{Additional arguments to be passed to graphical methods if needed.}
}
\description{
This function plots all the estimated functions (joint CDF, marginal survival and/or conditioncal cdf) from a \code{bivrecNP}() object in one step.
}
