% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bivrecNP.R
\name{plot.bivrecNP}
\alias{plot.bivrecNP}
\title{Plot Results of Non-Parametric Analysis of Bivariate Recurrent Events}
\usage{
\method{plot}{bivrecNP}(x, y = NULL, type = NULL, main = NULL,
  xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{bivrecNP}.}

\item{y}{either empty or NULL}

\item{type}{Optional vector of strings to label type 1and type 2 gap times. Default is c("Type 1", "Type 2").}

\item{main}{Optional string with plot title. Default is no title.}

\item{xlab}{Optional string with label for horizontal axis. Default is "Gap Times".}

\item{ylab}{Optional string with label for vertical axis. Default is "Individual".}

\item{...}{arguments to be passed to graphical methods as needed.}
}
\description{
This function plots the joint CDF, marginal survival and conditioncal cdf obtained from \code{bivrecNP}().
For examples see \code{bivrecNP}.
}
