% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bivrecReg.R
\name{plot.bivrecReg}
\alias{plot.bivrecReg}
\title{Plot Bivariate Alternating Recurrent Series from a bivrecReg object}
\usage{
\method{plot}{bivrecReg}(x, y = NULL, type = NULL, main = NULL,
  xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{bivrecReg}.}

\item{y}{Either empty or NULL}

\item{type}{Optional vector of strings to label type 1and type 2 gap times. Default is c("Type 1", "Type 2").}

\item{main}{Optional string with plot title. Default is no title.}

\item{xlab}{Optional string with label for horizontal axis. Default is "Gap Times".}

\item{ylab}{Optional string with label for vertical axis. Default is "Individual".}

\item{...}{arguments to be passed to graphical methods as needed.}
}
\description{
This function plots bivariate recurrent event gap times from a bivrecReg object. For examples see \code{bivrecReg}.
}
