% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.bivrecSurv}
\alias{plot.bivrecSurv}
\title{#' Produce Bivariate Alternating Recurrent Series Plot}
\usage{
\method{plot}{bivrecSurv}(object, main, xlab, ylab, type1, type2)
}
\arguments{
\item{object}{must be an object of \code{bivrecSurv}, \code{bivrecReg} or \code{bivrecNP} class or a formula that has a bivrecSurv class object as the response.}

\item{data}{a data frame that contains categorical covariates from a formula (if a formula is used for the object to be plotted).}
}
\description{
This function plots bivariate recurrent event gap times for either a bivrec object or a function. If a function is used plots will be produced for levels of categorical covariates.
}
\examples{
library(BivRec)

#Simulate data
set.seed(1234)
bivrec_data <- simulate(nsize=150, beta1=c(0.5,0.5), beta2=c(0,-0.5), tau_c=63, set=1.1)

#Plot a bivrecObject

bivrecSurvObject <- with(bivrec_data, bivrecSurv(id, epi, xij, yij, d1, d2))
plot(bivrecSurvObject)

#Plot a bivrecObject by categorical covariate from a formula
plot(bivrecSurv(id, epi, xij, yij, d1, d2) ~ a1 + a2, data = bivrec_data)

}
