% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSurv.R
\name{plot.bivrecSurv}
\alias{plot.bivrecSurv}
\title{#' Produce Bivariate Alternating Recurrent Series Plot}
\usage{
\method{plot}{bivrecSurv}(object, main, xlab, ylab, type1, type2)
}
\arguments{
\item{object}{must be an object of \code{bivrecSurv} class.}

\item{main}{optional string with plot title (default is no title)}

\item{xlab}{optional string with label for horizontal axis (default is "Gap Times")}

\item{ylab}{optional string with label for vertical axis (default is "Individual")}

\item{type1}{optional string to label type 1 gap times (default is "Type 1")}

\item{type2}{optional string to label type 2 gap times (default is "Type 2")}
}
\description{
This function plots bivariate recurrent event gap times from a bivrecSurv object.
}
\examples{
library(BivRec)

#Simulate data
set.seed(1234)
bivrec_data <- simulate(nsize=150, beta1=c(0.5,0.5), beta2=c(0,-0.5), tau_c=63, set=1.1)

#Create a bivrecSurv object
bivrecSurvObject <- with(bivrec_data, bivrecSurv(id, epi, xij, yij, d1, d2))

#Plot
plot(bivrecSurvObject)

}
