% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFormula.R
\name{plot.formula}
\alias{plot.formula}
\title{#' Produce Bivariate Alternating Recurrent Series Plots by Categories}
\usage{
\method{plot}{formula}(formula, data, main, xlab, ylab, type1, type2)
}
\arguments{
\item{formula}{must be a formula that has a bivrecSurv class object as the response.}

\item{data}{a data frame that contains categorical covariates and/or vectors to create the response indicate in the given formula.}

\item{main}{optional string with plot title (default is no title)}

\item{xlab}{optional string with label for horizontal axis (default is "Gap Times")}

\item{ylab}{optional string with label for vertical axis (default is "Individual")}

\item{type1}{optional string to label type 1 gap times (default is "Type 1")}

\item{type2}{optional string to label type 2 gap times (default is "Type 2")}
}
\description{
This function plots bivariate recurrent event gap times by levels of the categorical covariates as indicated in a formula.
}
\examples{
library(BivRec)

#Simulate data
set.seed(1234)
bivrec_data <- simulate(nsize=150, beta1=c(0.5,0.5), beta2=c(0,-0.5), tau_c=63, set=1.1)

#Plot by the categorical covariates in a formula
plot(bivrecSurv(id, epi, xij, yij, d1, d2) ~ a1 + a2, data = bivrec_data)

}
