% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bivrecNP.R
\name{plotCond}
\alias{plotCond}
\title{Plot Conditional CDF After Nonparametric Analysis}
\usage{
plotCond(object)
}
\arguments{
\item{object}{An object of \code{bivrecNP} class where the analysis has specified conditional = TRUE.}
}
\description{
This function plots conditional cdf for a \code{bivrecNP} object.
}
\keyword{internal}
