% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCond.R
\name{plotCond}
\alias{plotCond}
\title{Plot conditional CDF after non-parametric analysis}
\usage{
plotCond(x)
}
\arguments{
\item{x}{must be an object of \code{bivrecNP} class where the analysis has specified conditional = TRUE.}
}
\value{
A plot of conditional cdf in the given interval.
}
\description{
This function plots conditional cdf for recurrent event data.
}
\examples{
library(BivRec)
# Simulate bivariate alternating recurrent event data
set.seed(1234)
bivrec_data <- simulate(nsize=150, beta1=c(0.5,0.5), beta2=c(0,-0.5), tau_c=63, set=1.1)
bdat <- is.bivrecSurv(bivrec_data)
npresult <- bivrecNP(bdat,ai=1, u1 = c(2, 5, 10, 20), u2 = c(1, 5, 10, 15),conditional = TRUE, given.interval=c(0, 10))
plotCond(npresult)

}
