% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bivrecNP.R
\name{plotCond}
\alias{plotCond}
\title{Plot Conditional CDF After Nonparametric Analysis}
\usage{
plotCond(object, type)
}
\arguments{
\item{object}{An object of \code{bivrecNP} class where the analysis has specified conditional = TRUE.}

\item{type}{Optional string to label the Type II gap time. Default is "Type II".}
}
\description{
This function plots conditional cdf for a \code{bivrecNP} object.
}
