% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCond.R
\name{plotCond}
\alias{plotCond}
\title{Plot conditional CDF after non-parametric analysis}
\usage{
plotCond(x)
}
\arguments{
\item{BivrecNP}{object}
}
\value{
A plot of conditional cdf in given interval.
}
\description{
This function plots conditional cdf for recurrent event data.
}
\keyword{internal}
