% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCond.R
\name{plotCond}
\alias{plotCond}
\title{Plot conditional CDF after non-parametric analysis}
\usage{
plotCond(x)
}
\arguments{
\item{x}{must be an object of \code{bivrecNP} class where the analysis has specified conditional = TRUE.}
}
\value{
A plot of conditional cdf in the given interval.
}
\description{
This function plots conditional cdf for recurrent event data. For example see \code{bivrecNP}.
}
\keyword{internal}
