% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotJoint.R
\name{plotJoint}
\alias{plotJoint}
\title{Plot Joint CDF after non-parametric analysis}
\usage{
plotJoint(x)
}
\arguments{
\item{x}{must be an object of \code{bivrecNP} class.}
}
\value{
A contour plot of joint cdf.
}
\description{
This function plots joint cdf for recurrent event data. For example see \code{bivrecNP}.
}
\keyword{internal}
