% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bivrecNP.R
\name{plotJoint}
\alias{plotJoint}
\title{Plot Joint CDF After Nonparametric Analysis}
\usage{
plotJoint(object)
}
\arguments{
\item{object}{An object of \code{bivrecNP} class.}
}
\description{
This function plots the joint CDF for a \code{bivrecNP} object.
}
\keyword{internal}
