% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmarg.R
\name{plotMarg}
\alias{plotMarg}
\title{Plot Marginal Survival after non-parametric analysis}
\usage{
plotMarg(x, CI)
}
\arguments{
\item{x}{must be an object of \code{bivrecNP} class.}

\item{CI}{passed from object}
}
\value{
A plot of marginal survival vs. first gap time with confidence interval.
}
\description{
This function plots marginal survival for recurrent event data. For example see \code{bivrecNP}.
}
\keyword{internal}
