% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMethods.R
\name{print.bivrecNP}
\alias{print.bivrecNP}
\title{Print the Joint CDF, Marginal Survival, and Conditional CDF after Nonparametric Analysis}
\usage{
\method{print}{bivrecNP}(x, ...)
}
\arguments{
\item{x}{An object of class \code{bivrecNP}.}

\item{...}{Additional parameters if needed.}
}
\description{
This function prints a \code{bivrecNP} object.
}
