% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMethods.R
\name{print.bivrecNP}
\alias{print.bivrecNP}
\title{Print an Object of Class \code{bivrecNP}}
\usage{
\method{print}{bivrecNP}(x, ...)
}
\arguments{
\item{x}{An object of class \code{bivrecNP}.}

\item{...}{Additional parameters if needed.}
}
\description{
Print an Object of Class \code{bivrecNP}
}
