% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.bivrecNP}
\alias{print.bivrecNP}
\title{print}
\usage{
\method{print}{bivrecNP}(object)
}
\arguments{
\item{object}{of a bivrecNP object obtained by using the bivrecNP() function}
}
\description{
Print bivrecNP object
}
