% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_vcov_confint.R
\name{print.bivrecReg}
\alias{print.bivrecReg}
\title{Print an object of class bivrecReg}
\usage{
\method{print}{bivrecReg}(object)
}
\arguments{
\item{object}{An object of class bivrecReg}
}
\description{
Print an object of class bivrecReg
}
\keyword{internal}
