% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMethods.R
\name{print.bivrecReg}
\alias{print.bivrecReg}
\title{Print an Object of Class \code{bivrecReg}}
\usage{
\method{print}{bivrecReg}(x, ...)
}
\arguments{
\item{x}{An object of class \code{bivrecReg}.}

\item{...}{Additional parameters if needed.}
}
\description{
Print an Object of Class \code{bivrecReg}
}
