% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMethods.R
\name{print.summary.bivrecReg}
\alias{print.summary.bivrecReg}
\title{Print the Summary of a \code{bivrecReg} Object}
\usage{
\method{print}{summary.bivrecReg}(x, ...)
}
\arguments{
\item{x}{A \code{summary.bivrecReg} object.}

\item{...}{Additional parameters if needed.}
}
\description{
Print the Summary of a \code{bivrecReg} Object
}
