% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary.bivrecReg}
\alias{print.summary.bivrecReg}
\title{Print the summary of a bivrecReg object.}
\usage{
\method{print}{summary.bivrecReg}(x, ...)
}
\arguments{
\item{x}{a summary.bivrecReg object}

\item{...}{additional parameters if needed}
}
\description{
Print the summary of a bivrecReg object.
}
