% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semi.param.multivariate.R
\name{semi.param.multivariate}
\alias{semi.param.multivariate}
\title{A Function for multivariate fits using semiparametric regression method on a biv.rec object}
\usage{
semi.param.multivariate(new_data, cov_names, CI)
}
\arguments{
\item{new_data}{An object that has been reformatted for fit using the biv.rec.reformat() function. Passed from biv.rec.fit().}

\item{cov_names}{A vector with the names of the covariates. Passed from biv.rec.fit().}

\item{CI}{Logical. Passed from biv.rec.fit().}
}
\value{
A dataframe summarizing the estimates for effects of the covariates, their standard errors and 95% confidence intervals.
}
\description{
This function fits the semiparametric model given multiple  covariates. Called from biv.rec.fit(). No user interface.
}
\seealso{
\code{\link{biv.rec.fit}}
}
\keyword{internal}
