% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.bivrecReg}
\alias{summary.bivrecReg}
\title{summarize}
\usage{
\method{summary}{bivrecReg}(object)
}
\arguments{
\item{object}{a bivrecReg object obtained by using bivrecReg() function}
}
\description{
Summarize bivrecReg object
}
