% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.bivrecReg}
\alias{summary.bivrecReg}
\title{Summary of a bivrecReg object}
\usage{
\method{summary}{bivrecReg}(object, ...)
}
\arguments{
\item{object}{a bivrecReg object}

\item{...}{additional parameters if needed}
}
\description{
Summary of a bivrecReg object
}
