% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.bivrecReg}
\alias{summary.bivrecReg}
\title{Summary of a \code{bivrecReg} Object.}
\usage{
\method{summary}{bivrecReg}(object, ...)
}
\arguments{
\item{object}{A \code{bivrecReg} object}

\item{...}{Additional parameters if needed}
}
\description{
Summary of a \code{bivrecReg} Object.
}
