% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_vcov_confint.R
\name{vcov.bivrecReg}
\alias{vcov.bivrecReg}
\title{vcov}
\usage{
\method{vcov}{bivrecReg}(object, ...)
}
\arguments{
\item{object}{A bivrecReg object obtained by using bivrecReg() function}

\item{...}{additional parameters if needed}
}
\description{
Obtain variance-covariance matrix from semi-parametric regression fit using bivrecReg
}
