% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_vcov_confint.R
\name{vcov.bivrecReg}
\alias{vcov.bivrecReg}
\title{Extract the variance-covariance matrix of a semi-parametric regression fit from bivrecReg}
\usage{
\method{vcov}{bivrecReg}(object, ...)
}
\arguments{
\item{object}{A bivrecReg object}

\item{...}{additional parameters if needed}
}
\description{
Extract the variance-covariance matrix of a semi-parametric regression fit from bivrecReg
}
