% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_vcov_confint.R
\name{vcov.bivrecReg}
\alias{vcov.bivrecReg}
\title{Extract the Variance-Covariance Matrix of a Semiparametric Regression Fit from a \code{bivrecReg} object}
\usage{
\method{vcov}{bivrecReg}(object, ...)
}
\arguments{
\item{object}{A bivrecReg object.}

\item{...}{Additional parameters if needed.}
}
\description{
Extract the Variance-Covariance Matrix of a Semiparametric Regression Fit from a \code{bivrecReg} object
}
